+++ openjaus/trunk/OpenJAUSv3.3/libjaus/src/message/command/worldModel/terminateVksDataTransferMessage.c
@@ -490,6 +490,6 @@ static int headerToString(TerminateVksDataTransferMessage message, char **buf)
   strcat((*buf), "\nSequence Number: ");
   jausUnsignedShortToString(message->sequenceNumber, (*buf)+strlen(*buf));
   
-  return strlen((*buf));
+  return (int)strlen(*buf);
   
 }
